Update Detect and Recognize Car License Plate.py
It loads the pre-trained OCR model and its label file in load_graph and load_label functions. label_image_list method transforms the image to a tensor with the convert_tensor method and then predicts the label of the tensor with the label_image_list function and returns the license number. 
Code: Create a main function to perform the whole task in a sequence.